Optimize ui

+++ src/server/components/DebugConfiguration.ts
@@ -1,20 +1,22 @@
 import { IConfiguration, IMattermostConfigurationData } from "./IConfiguration";
 import { ITest } from "../../shared/ITest";
-import { IMattermost } from './IMattermost';
-import { DebugMattermost } from './DebugMattermost';
-import { InMemoryMongoDb } from './InMemoryMongoDb';
-import { IMongoDb } from './IMongoDb';
+import { IMattermost } from "./IMattermost";
+import { DebugMattermost } from "./DebugMattermost";
+import { InMemoryMongoDb } from "./InMemoryMongoDb";
+import { IMongoDb } from "./IMongoDb";
 import * as mongoose from "mongoose";
-import { MongoTestResult } from './MongoTestResult';
-import * as moment from 'moment';
-import { ITestResult } from '../../shared/ITestResult';
+import { MongoTestResult } from "./MongoTestResult";
+import * as moment from "moment";
+import { ITestResult } from "../../shared/ITestResult";
 
 export class DebugConfiguration implements IConfiguration {
+    private initialized: boolean = false;
+
     async tests(): Promise<ITest[]> {
         return [
             {
                 name: "google.de",
-                frequency: "10 seconds",
+                frequency: "30 seconds",
                 isActive: true,
                 type: "http-get",
                 options: {
@@ -24,7 +26,7 @@ export class DebugConfiguration implements IConfiguration {
             },
             {
                 name: "localhost demo service",
-                frequency: "10 seconds",
+                frequency: "30 seconds",
                 isActive: true,
                 type: "http-get",
                 options: {
@@ -41,41 +43,53 @@ export class DebugConfiguration implements IConfiguration {
 
     public async mongodb(): Promise<IMongoDb> {
         const inMemoryDb = new InMemoryMongoDb();
-        const running = await inMemoryDb.running();
-    
-        await mongoose.connect(running.url);
-        console.log('debug, connect');
 
+        if (this.initialized) return inMemoryDb;
+
+        console.log("Initializing demo data");
+
+        const running = await inMemoryDb.running();
+
+        await mongoose.connect(running.url);
         await MongoTestResult.deleteMany({});
-        console.log('debug, deletaMany');
 
         const today = moment();
 
-        const totalDays = 10;
+        const totalDays = 50;
         for (let index = 0; index < totalDays; index++) {
-            const currentDay = today.clone().subtract(index, 'days');
+            const currentDay = today.clone().subtract(index, "days");
 
-            const testsPerDay = Math.floor((Math.random() * 30) + 10);
+            const testsPerDay = Math.floor(Math.random() * 20 + 4);
             for (let test = 0; test < testsPerDay; test++) {
-                const successGoogle = Math.floor((Math.random() * 2) + 1) === 1;
+                const successGoogle = Math.floor(Math.random() * 2 + 1) === 1;
                 const testResultGoogle: ITestResult = {
-                    timestamp: currentDay.clone().startOf('day').add(test, 'minutes').toDate(),
-                    name: 'google.de',
+                    timestamp: currentDay
+                        .clone()
+                        .startOf("day")
+                        .add(test, "minutes")
+                        .toDate(),
+                    name: "google.de",
                     success: successGoogle,
                     data: { statusCode: successGoogle ? 200 : 500 }
                 };
                 await MongoTestResult.create(testResultGoogle);
-                
-                const successOther = Math.floor((Math.random() * 2) + 1) === 1;
+
+                const successOther = Math.floor(Math.random() * 2 + 1) === 1;
                 const testResultOther: ITestResult = {
-                    timestamp: currentDay.clone().startOf('day').add(test, 'minutes').toDate(),
-                    name: 'localhost demo service',
+                    timestamp: currentDay
+                        .clone()
+                        .startOf("day")
+                        .add(test, "minutes")
+                        .toDate(),
+                    name: "localhost demo service",
                     success: successOther,
                     data: { statusCode: successGoogle ? 200 : 500 }
                 };
-                await MongoTestResult.create(testResultOther);    
+                await MongoTestResult.create(testResultOther);
             }
         }
+
+        this.initialized = true;
 
         return inMemoryDb;
     }
